requirements.txt
requests>=2.26.0
pandas>=1.3.2
PyGithub>=1.55
argparse>=1.4.0
numpy>=1.19.5
nltk>=3.6.2
tensorflow>=2.6.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
opencv-python>=4.5.3.56
tf-explain>=0.3.0
pdoc3>=0.10.0